New Release + Optimization

+++ B-TREE/Program.cs
@@ -5,26 +5,27 @@ using System.IO;
 
 namespace B_TREE
 {
-    class Program
+    internal class Program
     {
-        static void Main(string[] args)
+        private static void Main(string[] args)
         {
-            BTree<int> tree = new BTree<int>(10);
+            BTree<int> tree = new BTree<int>(3);
 
-            tree.Load(@"test.json");
-            //for (int i = 0; i < 10000000; i++)
-            //{
-            //    tree.Add(i);
-            //}
-
-
-
+            //tree.Load(@"test.json");
+            for (int i = 0; i < 11; i++)
+            {
+                tree.Add(i);
+            }
+            tree.Save("tt.json");
+            var zz = new BTree<int>(3);
+            zz.Load("tt.json");
+            zz.Print();
 
             Console.WriteLine("Hello World!");
-            //for (int i = 0; i < 24; i++)
-            //{
-            //    Console.WriteLine(tree.Contains(i));
-            //}
+            for (int i = 0; i < 24; i++)
+            {
+                Console.WriteLine(tree.Contains(i));
+            }
             //tree.Print();
             //tree.Save(@"test.json");
             //tree.Print();
@@ -32,28 +33,19 @@ namespace B_TREE
         }
     }
 
-    public class BTree<T>:Node<T>
+    public class BTree<T> : Node<T>
     {
-        public Node<T> Root;
-        public new void Add(T value)
-        {
-            if(Root == null)
-            {
-                Root = new Node<T>(Order) {Owner=this, Leaf=true};
-            }
-            Root.Add(value);
-        }
-        public BTree(int order):base(order)
+        public BTree(int order) : base(order)
         {
             if (order < 3)
             {
                 throw new Exception("Order must be >2");
             }
-            Order = order;
         }
+
         public void Print()
         {
-            List<Node<T>> ForPrint = new List<Node<T>> { Root};
+            List<Node<T>> ForPrint = new List<Node<T>> { this };
             List<Node<T>> ForPrintNext = new List<Node<T>>();
             int i = 0;
             while (true)
@@ -61,16 +53,23 @@ namespace B_TREE
                 Console.WriteLine($"Leve: {i}");
                 foreach (var printNode in ForPrint)
                 {
-                    foreach (var item in printNode.Nodes)
+                    if (printNode.Nodes != null)
                     {
-                        ForPrintNext.Add(item);
+                        foreach (var item in printNode.Nodes)
+                        {
+                            ForPrintNext.Add(item);
+                        }
                     }
-                    string owner="";
-                    foreach (var item in printNode.Owner.Values)
+                    string owner = "";
+                    if (printNode.Owner != null)
                     {
-                        owner += item + " ";
+                        foreach (var item in printNode.Owner.Values)
+                        {
+                            owner += item + " ";
+                        }
                     }
-                    if(owner.Length!=0)
+
+                    if (owner.Length != 0)
                         owner = owner.Remove(owner.Length - 1);
 
                     Console.Write($"[{owner}](");
@@ -91,29 +90,7 @@ namespace B_TREE
                 i++;
             }
         }
-        public bool Contains(T value)
-        {
-            try
-            {
-                Root.Find(value);
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-        public new T Find(T value)
-        {
-            try
-            {
-                return Root.Find(value);
-            }
-            catch
-            {
-                return default(T);
-            }
-        }
+
         public void Save(string Path, Formatting formatting = Formatting.Indented)
         {
             using (StreamWriter file = File.CreateText(Path))
@@ -122,19 +99,23 @@ namespace B_TREE
                 {
                     Formatting = formatting
                 };
-                serializer.Serialize(file, Root);
+                serializer.Serialize(file, this);
             }
         }
+
         public void Load(string Path)
         {
-            Root = JsonConvert.DeserializeObject<Node<T>>(File.ReadAllText(Path));
-            Root.Owner = this;
-            Node<T> owner = Root;
-            OwnerRestore(Root);
+            Node<T> LoadedData = JsonConvert.DeserializeObject<Node<T>>(File.ReadAllText(Path));
+            Values = LoadedData.Values;
+            Nodes = LoadedData.Nodes;
+            Order = LoadedData.Order;
+            OwnerRestore(this);
         }
 
         private void OwnerRestore(Node<T> node)
         {
+            if (node.Nodes == null)
+                return;
             foreach (var subnode in node.Nodes)
             {
                 subnode.Owner = node;
@@ -142,27 +123,27 @@ namespace B_TREE
                 OwnerRestore(subnode);
             }
         }
-           
     }
 
     public class Node<T>
     {
         [JsonIgnore]
         public int Order { get; set; }
-        public bool Leaf { get; set; }
+
         public List<T> Values { get; set; }
         public List<Node<T>> Nodes { get; set; }
+
         [JsonIgnore]
         public Node<T> Owner { get; set; }
+
         public Node(int order)
         {
             Order = order;
-            Nodes = new List<Node<T>>();
-            Values = new List<T>();
         }
+
         public void Add(T value)
         {
-            if (!Leaf)
+            if (Nodes != null)
             {
                 int i;
                 for (i = 0; i < Values.Count; i++)
@@ -178,66 +159,57 @@ namespace B_TREE
             else
             {
                 int i;
+                if (Values == null)
+                {
+                    Values = new List<T>();
+                }
                 for (i = 0; i < Values.Count; i++)
                 {
                     if (Comparer<T>.Default.Compare(value, Values[i]) != 1)
                     {
                         Values.Insert(i, value);
-                        LeafCheck();
+                        Check();
                         return;
                     }
                 }
                 Values.Insert(i, value);
-                LeafCheck();
+                Check();
             }
         }
-        private void LeafCheck()
+
+        private void Check()
         {
             if (Values.Count == Order)
             {
-                List<T> NewElms = new List<T>(Values.GetRange(Order / 2 + 1, Order - (Order / 2 + 1)));
-                Node<T> NewNode = new Node<T>(Order) { Values = NewElms, Leaf = true };
+                Node<T> NewNode = new Node<T>(Order);
                 T ValueToInsert = Values[Order / 2];
+                List<T> NewElms = new List<T>(Values.GetRange(Order / 2 + 1, Order - (Order / 2 + 1)));
+                NewNode.Values = NewElms;
                 Values.RemoveRange(Order / 2, Order - (Order / 2));
-                if (Owner is BTree<T>)
+
+                if (Nodes != null)
                 {
-                    BTree<T> tree = Owner as BTree<T>;
-                    Node<T> NewOwner = new Node<T>(Order) {Owner = tree, Nodes =new List<Node<T>> {this,NewNode },Values=new List<T> { ValueToInsert } };
-                    NewNode.Owner = NewOwner;
-                    tree.Root = NewOwner;
-                    Owner = NewOwner;
+                    List<Node<T>> NewNodes = new List<Node<T>>(Nodes.GetRange(Order / 2 + 1, Order - (Order / 2)));
+                    foreach (var node in NewNodes)
+                    {
+                        node.Owner = NewNode;
+                    }
+                    NewNode.Nodes = NewNodes;
+                    Nodes.RemoveRange(Order / 2 + 1, Order - (Order / 2));
+                }
+                if (Owner == null)
+                {
+                    NewNode.Owner = this;
+                    Node<T> NewNode2 = new Node<T>(Order) { Owner = this, Nodes = Nodes, Values = Values };
+                    Values = new List<T> { ValueToInsert };
+                    Nodes = new List<Node<T>> { NewNode2, NewNode };
                     return;
                 }
                 Owner.Insert(ValueToInsert, NewNode);
             }
         }
-        private void NodeCheck()
-        {
-            if (Values.Count == Order)
-            {
-                List<T> NewElms = new List<T>(Values.GetRange(Order / 2 + 1, Order - (Order / 2 + 1)));
-                List<Node<T>> NewNodes = new List<Node<T>>(Nodes.GetRange(Order / 2+1, Order - (Order / 2)));
-                Node<T> NewNode = new Node<T>(Order) { Nodes = NewNodes, Values = NewElms, Leaf = false };
-                foreach (var node in NewNodes)
-                {
-                    node.Owner = NewNode;
-                }
-                T ValueToInsert = Values[Order / 2];
-                Values.RemoveRange(Order / 2, Order - (Order / 2));
-                Nodes.RemoveRange(Order / 2+1, Order - (Order / 2));
-                if (Owner is BTree<T>)
-                {
-                    BTree<T> tree = Owner as BTree<T>;
-                    Node<T> NewOwner = new Node<T>(Order) { Owner = tree, Nodes = new List<Node<T>> { this,NewNode }, Values = new List<T> { ValueToInsert } };
-                    NewNode.Owner = NewOwner;
-                    tree.Root = NewOwner;
-                    Owner = NewOwner;
-                    return;
-                }
-                Owner.Insert(ValueToInsert, NewNode);
-            }
-        }
-        private void Insert(T value,Node<T> nodeValue)
+
+        private void Insert(T value, Node<T> nodeValue)
         {
             nodeValue.Owner = this;
             int i;
@@ -246,38 +218,39 @@ namespace B_TREE
                 if (Comparer<T>.Default.Compare(value, Values[i]) != 1)
                 {
                     Values.Insert(i, value);
-                    Nodes.Insert(i+1, nodeValue);
-                    NodeCheck();
+                    Nodes.Insert(i + 1, nodeValue);
+                    Check();
                     return;
                 }
             }
             Values.Insert(i, value);
-            Nodes.Insert(i+1, nodeValue);
-            NodeCheck();
+            Nodes.Insert(i + 1, nodeValue);
+            Check();
         }
-        public T Find(T value)
+
+        public bool Contains(T value)
         {
             int i;
             for (i = 0; i < Values.Count; i++)
             {
                 if (Comparer<T>.Default.Compare(Values[i], value) == 0)
                 {
-                    return Values[i];
+                    return true;
                 }
                 if (Comparer<T>.Default.Compare(Values[i], value) == 1)
                 {
-                    if (Leaf)
+                    if (Nodes == null)
                     {
-                        throw new Exception("Not Found");
+                        return false;
                     }
-                    return Nodes[i].Find(value);
+                    return Nodes[i].Contains(value);
                 }
             }
-            if (Leaf)
+            if (Nodes == null)
             {
-                throw new Exception("Not Found");
+                return false;
             }
-            return Nodes[i].Find(value);
+            return Nodes[i].Contains(value);
         }
     }
 }
